requirements.txt
torch
torchvision
torchaudio
transformer
datasets
cython
imageio
multipledispatch==0.6.0
git+https://github.com/facebookresearch/detectron2.git
opencv-python
scikit-learn
tqdm
python-doctr[torch]